chore: move defaultTranslatorFunction in /i18n/Streami18n to /i18n/utils
