+++ src/components/MediaRecorder/classes/MediaRecorderController.ts
@@ -16,7 +16,7 @@ import {
   RecordedMediaType,
 } from '../../ReactFileUtilities';
 import { TranslationContextValue } from '../../../context/TranslationContext';
-import { defaultTranslatorFunction } from '../../../i18n';
+import { defaultTranslatorFunction } from '../../../i18n/utils';
 import { isSafari } from '../../../utils/browsers';
 import { mergeDeepUndefined } from '../../../utils/mergeDeep';
